nios.c: compiles cleanly, moving on to osal.c

+++ lsf/include/lib/nios.h
@@ -11,4 +11,4 @@ int ls_stoprex(void);
 int ls_niossync(long numTasks);
 int ls_setstdout(int on_, char *format);
 int ls_setstdin(int on_, pid_t *rpidlist, size_t len);
-long ls_getstdin(int on_, pid_t *rpidlist, long maxlen);
+size_t ls_getstdin(int on_, pid_t *rpidlist, long maxlen);
